soft spi code. TODO : fix pin mappings

+++ ESP8266_SPI.h
@@ -22,7 +22,7 @@
 
 void ESP8266_SPI_init_pins(void);
 void ESP8266_SPI_set_params(void);
-void ESP8266_SPI_send(uint32_t address, uint32_t data);
+void ESP8266_SPI_send(uint8_t address, uint8_t data);
 void ESP8266_SPI_get(uint8_t* data);
 
 #endif
